requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.27.2
pydantic==2.11.7
pydantic-settings==2.10.1 # Added for BaseSettings
SQLAlchemy==2.0.42
psycopg[binary]==3.2.9
redis==6.2.0
celery==5.5.3
alembic==1.16.4
prometheus-client==0.22.1
pyyaml==6.0.2
python-json-logger==3.3.0
python-dotenv==1.1.1
openapi-python-client==0.22.0
authlib==1.4.0
itsdangerous==2.2.0

# Development dependencies (also needed for CI)
flake8==7.3.0
black==24.10.0
mypy==1.10.1
types-PyYAML==6.0.12.20250516